backend/grocery/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-richtextfield==1.6.1
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
